Implement the navigation to Movie Details screen

+++ feature/tmdb/api/src/main/java/com/majorkik/tmdb/api/repository/MovieDetailsRepository.kt
@@ -4,5 +4,5 @@ import com.majorkik.tmdb.api.model.MovieDetails
 import com.majorkik.tmdb.api.network.NetworkResult
 
 interface MovieDetailsRepository {
-    suspend fun fetchMovieDetails(id: Long): NetworkResult<MovieDetails, String>
+    suspend fun fetchMovieDetails(id: Int): NetworkResult<MovieDetails, String>
 }
